refactor: J-Quants client を provider adapter 配下へ移動
market data provider 抽象化に向けて、JQuantsClient を services/market_data/providers/jquants.rs に移動する。API path と response shape は維持する。
