+++ backend/src/services.rs
@@ -7,7 +7,7 @@ pub mod csv_util;
 pub mod dividend;
 pub mod dividend_cache;
 pub mod domestic_stock;
-pub mod jquants;
+pub mod market_data;
 pub mod mutualfund;
 pub mod shared;
 pub mod stock;
