+++ backend/src/services/market_data/mod.rs
@@ -1,0 +1,1 @@
+pub mod providers;
